build_python_mango/requirement.txt
Flask
pymongo
kafka-python
six
